+++ Cargo.toml
@@ -29,19 +29,8 @@ anyhow = "1.0.98"
 base64ct = { version = "1.8.0", features = ["std"] }
 credibil-core = { git = "https://github.com/credibil/vdc.git" }
 credibil-ecc = { path = "ecc" }
-# credibil-keyring = { path = "crates/keyring" }
-# credibil-kms = { path = "crates/kms" }
-ecies = { version = "0.2.9", default-features = false, features = ["pure", "std"] }
-ed25519-dalek = { version = "2.1.1", features = ["rand_core"] }
-multibase = "0.9.1"
-rand = "0.8.5"
 serde = { version = "1.0.219", features = ["derive"] }
-serde_json = { version = "1.0.140", features = ["alloc"] }
-sha2 = "0.10.9"
-# test-kms = { path = "crates/test-kms" }
 tokio = { version = "1.45.1", features = ["macros", "rt-multi-thread"] }
-tracing = "0.1.41"
-x25519-dalek = { version = "2.0.1", features = ["static_secrets"] }
 zeroize = "1.8.1"
 
 [patch."https://github.com/credibil/vdc.git"]
