+++ ecc/Cargo.toml
@@ -28,17 +28,15 @@ chacha20poly1305 = "0.10.1"
 ciborium = "0.2.2"
 credibil-core.workspace = true
 ecdsa = "0.16.9"
-ecies.workspace = true
-ed25519-dalek.workspace = true
+ecies = { version = "0.2.9", default-features = false, features = ["pure", "std"] }
+ed25519-dalek = { version = "2.1.1", features = ["rand_core"] }
 k256 = "0.13.4"
 p256 = "0.13.2"
-rand.workspace = true
+rand = "0.8.5"
 serde.workspace = true
-sha2.workspace = true
-x25519-dalek.workspace = true
+sha2 = "0.10.9"
+x25519-dalek = { version = "2.0.1", features = ["static_secrets"] }
 zeroize.workspace = true
 
 [dev-dependencies]
-hex = "0.4.3"
-signature = "2.2.0"
 tokio.workspace = true
